Add base class for path planner nodes (normal one and controller test node) to store shared code
